@@ -32,10 +32,10 @@
   "homepage": "https://github.com/jcarpenter/gambier#readme",
   "dependencies": {
     "@rollup/plugin-json": "^4.1.0",
-    "better-sqlite3": "^7.1.1",
+    "better-sqlite3": "^7.5.0",
     "bezier-easing": "^2.1.0",
-    "chokidar": "^3.4.3",
-    "chroma-js": "^2.1.0",
+    "chokidar": "^3.5.3",
+    "chroma-js": "^2.4.2",
     "citeproc": "^2.4.48",
     "codemirror-markdown-list-autoindent": "^1.0.1",
     "color": "^3.1.3",
@@ -46,7 +46,6 @@
     "deep-eql": "^4.0.0",
     "deep-object-diff": "^1.1.0",
     "electron-localshortcut": "^3.2.1",
-    "electron-store": "^6.0.1",
     "fs-extra": "^9.0.1",
     "gray-matter": "^4.0.2",
     "hierarchy-js": "^1.0.4",
@@ -60,9 +59,9 @@
     "path": "^0.12.7",
     "remove-markdown": "^0.3.0",
     "sqlite3": "^5.0.0",
-    "svelte": "^3.30.1",
-    "svelte-resize-observer": "^1.1.0",
-    "turndown": "^7.0.0",
+    "svelte": "^3.46.6",
+    "svelte-resize-observer": "^2.0.0",
+    "turndown": "^7.1.1",
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
@@ -70,16 +69,17 @@
     "@rollup/plugin-node-resolve": "^11.0.0",
     "codemirror": "^5.58.2",
     "cpy-cli": "^3.1.1",
-    "electron": "^12.0.2",
+    "electron": "^18.0.1",
     "electron-builder": "^22.9.1",
     "electron-rebuild": "^2.3.4",
     "electron-reload": "^1.5.0",
+    "electron-store": "^8.0.1",
     "npm-run-all": "^4.1.5",
     "rollup": "^2.34.0",
     "rollup-plugin-svelte": "^6.1.1",
-    "sass": "^1.29.0",
-    "svelte-check": "^1.1.28",
-    "svelte-preprocess": "^4.6.1",
-    "svgo": "^2.3.0"
+    "sass": "^1.49.10",
+    "svelte-check": "^2.4.6",
+    "svelte-preprocess": "^4.10.4",
+    "svgo": "^2.8.0"
   }
 }
