@@ -52,7 +52,7 @@
 
 	.excerpt {
 		@include system-regular-font;
-		color: var(--secondary-label-color);
+		color: var(--label-2-color);
 		display: -webkit-box;
 		-webkit-box-orient: vertical;
 		-webkit-line-clamp: 3;
@@ -82,7 +82,7 @@
 		height: 79px; // height minus 1, to create 1px gap below
 		margin-bottom: 1px;
 		.title, .excerpt {
-			color: var(--selected-menuitem-text-color);
+			color: var(--menuitem-selected-text-color);
 		}
 		&.isSidebarFocused {
 		  background-color: var(--selected-list-item-color);
@@ -90,7 +90,7 @@
 	
 		// Selected, and parent list NOT focused
 		&:not(.isSidebarFocused) {
-		  background-color: var(--disabledControlTextColor);
+		  background-color: var(--control-disabled-text-color);
 		}
 	}
 </style>
